# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VmStatusState = typeof VmStatusState[keyof typeof VmStatusState];


export const VmStatusState = {
  stopped: 'stopped',
  starting: 'starting',
  running: 'running',
  stopping: 'stopping',
  error: 'error',
} as const;

export type VmStatusConnectionMode = typeof VmStatusConnectionMode[keyof typeof VmStatusConnectionMode];


export const VmStatusConnectionMode = {
  serial: 'serial',
  ssh: 'ssh',
} as const;

export interface VmStatus {
  state: VmStatusState;
  /** @nullable */
  pid?: number | null;
  /**
     * Uptime in seconds
     * @nullable
     */
  uptime?: number | null;
  /** @nullable */
  isoPath?: string | null;
  /** @nullable */
  diskPath?: string | null;
  ramGb: number;
  cpuCores: number;
  /** @nullable */
  gpuPassthrough?: string | null;
  connectionMode: VmStatusConnectionMode;
  /** @nullable */
  sshPort?: number | null;
}

export interface VmActionResult {
  success: boolean;
  message: string;
  state: string;
}

export type VmConfigConnectionMode = typeof VmConfigConnectionMode[keyof typeof VmConfigConnectionMode];


export const VmConfigConnectionMode = {
  serial: 'serial',
  ssh: 'ssh',
} as const;

export interface VmConfig {
  /** @nullable */
  isoPath?: string | null;
  /** @nullable */
  diskPath?: string | null;
  ramGb: number;
  cpuCores: number;
  /** @nullable */
  gpuPassthrough?: string | null;
  connectionMode: VmConfigConnectionMode;
  sshPort: number;
  /** @nullable */
  sshUser?: string | null;
  /** @nullable */
  sshPassword?: string | null;
}

export type VmConfigUpdateConnectionMode = typeof VmConfigUpdateConnectionMode[keyof typeof VmConfigUpdateConnectionMode];


export const VmConfigUpdateConnectionMode = {
  serial: 'serial',
  ssh: 'ssh',
} as const;

export interface VmConfigUpdate {
  /** @nullable */
  isoPath?: string | null;
  /** @nullable */
  diskPath?: string | null;
  ramGb?: number;
  cpuCores?: number;
  /** @nullable */
  gpuPassthrough?: string | null;
  connectionMode?: VmConfigUpdateConnectionMode;
  sshPort?: number;
  /** @nullable */
  sshUser?: string | null;
  /** @nullable */
  sshPassword?: string | null;
}

export interface SnapshotInput {
  name: string;
}

export interface ShellCommandInput {
  command: string;
  /** Max ms to wait for output (default 10000) */
  timeoutMs?: number;
}

export interface ShellCommandResult {
  stdout: string;
  stderr: string;
  /** @nullable */
  exitCode?: number | null;
  timedOut: boolean;
}

export interface ShellHistoryEntry {
  id: number;
  command: string;
  stdout: string;
  stderr: string;
  /** @nullable */
  exitCode?: number | null;
  timestamp: string;
}

export type FileEntryType = typeof FileEntryType[keyof typeof FileEntryType];


export const FileEntryType = {
  file: 'file',
  directory: 'directory',
  symlink: 'symlink',
  other: 'other',
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  /** @nullable */
  sizeBytes?: number | null;
  /** @nullable */
  modifiedMs?: number | null;
}

export interface DirectoryListing {
  path: string;
  /** @nullable */
  parent?: string | null;
  entries: FileEntry[];
}

export interface DriveInfo {
  name: string;
  path: string;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  fsType?: string | null;
  /** @nullable */
  sizeBytes?: number | null;
  removable: boolean;
}

export type FrontloadInputCategory = typeof FrontloadInputCategory[keyof typeof FrontloadInputCategory];


export const FrontloadInputCategory = {
  drivers: 'drivers',
  isos: 'isos',
  files: 'files',
} as const;

export interface FrontloadInput {
  sources: string[];
  category?: FrontloadInputCategory;
}

export interface CopiedItem {
  source: string;
  destination: string;
}

export interface FrontloadFailure {
  source: string;
  error: string;
}

export interface FrontloadResult {
  stagingPath: string;
  copied: CopiedItem[];
  failed: FrontloadFailure[];
}

export type ListDirectoryParams = {
/**
 * Absolute directory path to list (defaults to home)
 */
path?: string;
};