# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odysseus VM Shell API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vm/snapshot:
    post:
      operationId: snapshotVm
      tags: [vm]
      summary: Take a VM snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotInput"
      responses:
        "200":
          description: Snapshot taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmActionResult"
  /vm/config:
    get:
      operationId: getVmConfig
      tags: [vm]
      summary: Get VM hardware configuration
      responses:
        "200":
          description: VM hardware configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmConfig"
    put:
      operationId: updateVmConfig
      tags: [vm]
      summary: Update VM hardware configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VmConfigUpdate"
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmConfig"
  /shell/exec:
    post:
      operationId: execShellCommand
      tags: [shell]
      summary: Execute a command in the shell (for Odysseus agent tool)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShellCommandInput"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShellCommandResult"
...
  /files/list:
    get:
      operationId: listDirectory
      tags: [files]
      summary: List entries in a directory
      parameters:
        - name: path
          in: query
          required: false
          description: Absolute directory path to list (defaults to home)
          schema:
            type: string
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryListing"
...
  /files/frontload:
    post:
      operationId: frontloadFiles
      tags: [files]
      summary: Copy files/drivers from a source (USB) into the staging area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrontloadInput"
      responses:
        "200":
          description: Frontload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrontloadResult"
components:
  schemas:
...
    VmActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        state:
          type: string
      required:
        - success
        - message
        - state
    VmConfig:
      type: object
      properties:
        isoPath:
          type: ["string", "null"]
        diskPath:
          type: ["string", "null"]
        ramGb:
          type: integer
        cpuCores:
          type: integer
        gpuPassthrough:
          type: ["string", "null"]
        connectionMode:
          type: string
          enum: [serial, ssh]
        sshPort:
          type: integer
        sshUser:
          type: ["string", "null"]
        sshPassword:
          type: ["string", "null"]
      required:
        - ramGb
        - cpuCores
        - connectionMode
        - sshPort
    VmConfigUpdate:
      type: object
      properties:
        isoPath:
          type: ["string", "null"]
        diskPath:
          type: ["string", "null"]
        ramGb:
          type: integer
        cpuCores:
          type: integer
        gpuPassthrough:
          type: ["string", "null"]
        connectionMode:
          type: string
          enum: [serial, ssh]
        sshPort:
          type: integer
        sshUser:
          type: ["string", "null"]
        sshPassword:
          type: ["string", "null"]
    SnapshotInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ShellCommandInput:
      type: object
      properties:
        command:
          type: string
        timeoutMs:
          type: integer
          description: Max ms to wait for output (default 10000)
      required:
        - command
    ShellCommandResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: ["integer", "null"]
        timedOut:
          type: boolean
      required:
        - stdout
        - stderr
        - timedOut
...
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink, other]
        sizeBytes:
          type: ["integer", "null"]
        modifiedMs:
          type: ["integer", "null"]
      required:
        - name
        - path
        - type
    DirectoryListing:
      type: object
      properties:
        path:
          type: string
        parent:
          type: ["string", "null"]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
      required:
        - path
        - entries
...
    FrontloadInput:
      type: object
      properties:
        sources:
          type: array
          items:
            type: string
        category:
          type: string
          enum: [drivers, isos, files]
      required:
        - sources
    CopiedItem:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
      required:
        - source
        - destination
    FrontloadFailure:
      type: object
      properties:
        source:
          type: string
        error:
          type: string
      required:
        - source
        - error
    FrontloadResult:
      type: object
      properties:
        stagingPath:
          type: string
        copied:
          type: array
          items:
            $ref: "#/components/schemas/CopiedItem"
        failed:
          type: array
          items:
            $ref: "#/components/schemas/FrontloadFailure"
      required:
        - stagingPath
        - copied
        - failed