# lib/api-zod/src/generated/types/copiedItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */

export interface CopiedItem {
  source: string;
  destination: string;
}