# lib/api-zod/src/generated/types/directoryListing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntry } from './fileEntry';

export interface DirectoryListing {
  path: string;
  /** @nullable */
  parent?: string | null;
  entries: FileEntry[];
}