# lib/api-zod/src/generated/types/fileEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntryType } from './fileEntryType';

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  /** @nullable */
  sizeBytes?: number | null;
  /** @nullable */
  modifiedMs?: number | null;
}