# lib/api-zod/src/generated/types/frontloadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */
import type { FrontloadInputCategory } from './frontloadInputCategory';

export interface FrontloadInput {
  sources: string[];
  category?: FrontloadInputCategory;
}