# lib/api-zod/src/generated/types/frontloadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Odysseus VM Shell API
 * OpenAPI spec version: 0.1.0
 */
import type { CopiedItem } from './copiedItem';
import type { FrontloadFailure } from './frontloadFailure';

export interface FrontloadResult {
  stagingPath: string;
  copied: CopiedItem[];
  failed: FrontloadFailure[];
}